requirements.txt
APScheduler==3.8.1
backports.zoneinfo==0.2.1
click==8.0.3
cycler==0.11.0
dataclasses==0.8
Flask==1.1.0
flask-mongoengine==0.9.1
Flask-Script==1.0.0
Flask-WTF==0.14.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
mongoengine==0.23.1
numpy==1.19.5
Pillow==8.4.0
pkg-resources==0.0.0
pycryptodome==3.12.0
pymongo==3.12.3
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
typing-extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
Werkzeug==2.0.2
WTForms==3.0.0
zipp==3.6.0
